+++ UnitTests/dotNetInstallerToolsLibUnitTests/FindWindow.h
@@ -1,0 +1,14 @@
+#pragma once
+
+namespace DVLib
+{
+	namespace UnitTests 
+	{
+		class FindWindow
+		{
+		public:
+			HWND static FindWindowFromProcessId( DWORD dwProcessId );
+			HWND static FindWindowFromProcess( HANDLE hProcess );
+		};
+	}
+}